// utils/types.ts
export type EventType = {
    icon: StaticImageData
    heading: string;
    content: string;
}

export type PersonType = {
    image: StaticImageData;
    name: string;
    position: string;
    feedback: string;
}